+++ cmd/sfreleaser/main.go
@@ -21,8 +21,9 @@ import (
 var version = "dev"
 
 var (
-	devSkipGoreleaser       = os.Getenv("SFRELEASER_DEV_SKIP_GORELEASER") == "true"
-	devSkipRustCargoPublish = os.Getenv("SFRELEASER_DEV_SKIP_RUST_CARGO_PUBLISH") == "true"
+	devSkipGoreleaser                 = os.Getenv("SFRELEASER_DEV_SKIP_GORELEASER") == "true"
+	devSkipRustCargoPublish           = os.Getenv("SFRELEASER_DEV_SKIP_RUST_CARGO_PUBLISH") == "true"
+	devSkipSubstreamsRegistryPublish  = os.Getenv("SFRELEASER_DEV_SKIP_SUBSTREAMS_REGISTRY_PUBLISH") == "true"
 )
 
 var (