+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## v0.11.0
+
+- Initial release for supporting Substreams variant under language Rust.
+
+  Requires latest Substreams CLI for correctly publish packages to the registry.
+
 ## v0.10.1
 
 - Fixed `sfreleaser build` not using correct latest `goreleaser/goreleaser-cross:v1.24` image.